call: Rename files to follow the prefix-name rule
Renames the call files to xdp-call.
