+++ src/gamemode.c
@@ -22,7 +22,7 @@
 
 #include "config.h"
 
-#include "call.h"
+#include "xdp-call.h"
 #include "permissions.h"
 
 #include "xdp-app-info.h"
