+++ src/proxy-resolver.c
@@ -26,7 +26,7 @@
 #include <gio/gio.h>
 
 #include "proxy-resolver.h"
-#include "call.h"
+#include "xdp-call.h"
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
