+++ src/realtime.c
@@ -26,7 +26,7 @@
 #include <gio/gio.h>
 
 #include "realtime.h"
-#include "call.h"
+#include "xdp-call.h"
 #include "permissions.h"
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
