+++ src/session.c
@@ -20,7 +20,7 @@
 
 #include "session.h"
 #include "request.h"
-#include "call.h"
+#include "xdp-call.h"
 
 #include <string.h>
