+++ src/session.h
@@ -21,7 +21,7 @@
 #pragma once
 
 #include "request.h"
-#include "call.h"
+#include "xdp-call.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
