+++ src/trash.c
@@ -35,7 +35,7 @@
 #include <gio/gunixfdlist.h>
 
 #include "trash.h"
-#include "call.h"
+#include "xdp-call.h"
 #include "documents.h"
 #include "xdp-app-info.h"
 #include "xdp-dbus.h"
